+++ api/system/node/api_views.py
@@ -4,20 +4,18 @@ from django.conf import settings
 from django.utils.six import text_type
 
 from api.api_views import APIView
-from api.exceptions import (NodeIsNotOperational, PreconditionRequired, TaskIsAlreadyRunning,
-                            ObjectNotFound, GatewayTimeout)
+# noinspection PyProtectedMember
+from api.fields import get_boolean_value
+from api.exceptions import NodeIsNotOperational, PreconditionRequired, ObjectNotFound, GatewayTimeout
 from api.node.utils import get_node, get_nodes
 from api.system.messages import LOG_SYSTEM_UPDATE
 from api.system.node.serializers import NodeVersionSerializer
-from api.system.node.events import NodeUpdateStarted, NodeUpdateFinished
 from api.system.service.control import NodeServiceControl
 from api.system.update.serializers import UpdateSerializer
-from api.system.update.utils import process_update_reply
-from api.task.response import SuccessTaskResponse, FailureTaskResponse
-from que import TG_DC_UNBOUND, TT_DUMMY, Q_FAST
-from que.lock import TaskLock
-from que.utils import task_id_from_request, worker_command
-from vms.models import DefaultDc
+from api.task.response import SuccessTaskResponse, FailureTaskResponse, TaskResponse
+from que import Q_FAST, TG_DC_UNBOUND
+from que.utils import worker_command
+from que.tasks import execute
 
 
 logger = getLogger(__name__)
@@ -34,6 +32,9 @@ class NodeVersionView(APIView):
 
         if hostname:
             self.node = get_node(request, hostname)
+
+            if self.data and get_boolean_value(self.data.get('fresh', None)):
+                del self.node.system_version  # Remove cached version information
         else:
             self.node = get_nodes(request)
 
@@ -70,28 +71,37 @@ class NodeServiceStatusView(APIView):
 
 class NodeUpdateView(APIView):
     """api.system.node.views.system_node_update"""
+    LOCK = 'system_node_update:%s'
     dc_bound = False
-    _lock_key = 'system_update'
 
     def __init__(self, request, hostname, data):
         super(NodeUpdateView, self).__init__(request)
         self.hostname = hostname
         self.data = data
         self.node = get_node(request, hostname)
-        self.task_id = task_id_from_request(self.request, dummy=True, tt=TT_DUMMY, tg=TG_DC_UNBOUND)
 
-    def _update(self, version, key=None, cert=None):
+    def _update_v2(self, version, key=None, cert=None):
+        from api.system.update.utils import process_update_reply
+
         node = self.node
         worker = node.worker(Q_FAST)
-        logger.debug('Running node "%s" system update to version: "%s"', node, version)
+        logger.info('Running oldstyle (v2.x) node "%s" system update to version: "%s"', node, version)
         reply = worker_command('system_update', worker, version=version, key=key, cert=cert, timeout=600)
 
         if reply is None:
             raise GatewayTimeout('Node worker is not responding')
 
-        response_class, result = process_update_reply(reply, node, version)
-        response = response_class(self.request, result, task_id=self.task_id, obj=node, msg=LOG_SYSTEM_UPDATE,
-                                  detail_dict=result, dc_bound=False)
+        result, error = process_update_reply(reply, node, version)
+
+        if error:
+            response_class = FailureTaskResponse
+        else:
+            response_class = SuccessTaskResponse
+
+        detail_dict = result.copy()
+        detail_dict['version'] = version
+        response = response_class(self.request, result, obj=node, msg=LOG_SYSTEM_UPDATE, dc_bound=False,
+                                  detail_dict=detail_dict)
 
         if response.status_code == 200:
             # Restart all erigonesd workers
@@ -102,49 +112,72 @@ class NodeUpdateView(APIView):
 
         return response
 
-    @classmethod
-    def get_task_lock(cls):
-        # Also used in socket.io namespace
-        return TaskLock(cls._lock_key, desc='System task')
-
     def put(self):
-        assert self.request.dc.id == DefaultDc().id
+        assert self.request.dc.is_default()
 
         ser = UpdateSerializer(self.request, data=self.data)
 
         if not ser.is_valid():
-            return FailureTaskResponse(self.request, ser.errors, task_id=self.task_id, dc_bound=False)
+            return FailureTaskResponse(self.request, ser.errors, dc_bound=False)
 
         node = self.node
-        version = ser.object['version']
-
+        version = ser.data['version']
+        key = ser.data.get('key')
+        cert = ser.data.get('cert')
+        del node.system_version  # Request latest version in next command
         node_version = node.system_version
 
         if not (isinstance(node_version, text_type) and node_version):
             raise NodeIsNotOperational('Node version information could not be retrieved')
 
-        if version == ('v' + node.system_version):
+        node_version = node_version.split(':')[-1]  # remove edition prefix
+
+        if version == ('v' + node_version) and not ser.data.get('force'):
             raise PreconditionRequired('Node is already up-to-date')
 
         if node.status != node.OFFLINE:
-            raise NodeIsNotOperational('Unable to perform update on node that is not in maintenance state!')
+            raise NodeIsNotOperational('Unable to perform update on node that is not in maintenance state')
 
-        lock = self.get_task_lock()
+        if node_version.startswith('2.'):
+            # Old-style (pre 3.0) update mechanism
+            return self._update_v2(version, key=key, cert=cert)
 
-        if not lock.acquire(self.task_id, timeout=7200, save_reverse=False):
-            raise TaskIsAlreadyRunning
+        # Upload key and cert and get command array
+        worker = node.worker(Q_FAST)
+        update_cmd = worker_command('system_update_command', worker, version=version, key=key, cert=cert,
+                                    force=ser.data.get('force'), timeout=10)
 
-        try:
-            # Emit event into socket.io
-            NodeUpdateStarted(self.task_id, request=self.request).send()
+        if update_cmd is None:
+            raise GatewayTimeout('Node worker is not responding')
 
-            return self._update(version, key=ser.object.get('key'), cert=ser.object.get('cert'))
-        finally:
-            lock.delete(fail_silently=True, delete_reverse=False)
-            # Delete cached node version information (will be cached again during next node.system_version call)
-            del node.system_version
-            # Emit event into socket.io
-            NodeUpdateFinished(self.task_id, request=self.request).send()
+        if not isinstance(update_cmd, list):
+            raise PreconditionRequired('Node update command could be retrieved')
+
+        msg = LOG_SYSTEM_UPDATE
+        _apiview_ = {
+            'view': 'system_node_update',
+            'method': self.request.method,
+            'hostname': node.hostname,
+            'version': version,
+        }
+        meta = {
+            'apiview': _apiview_,
+            'msg': msg,
+            'node_uuid': node.uuid,
+            'output': {'returncode': 'returncode', 'stdout': 'message'},
+            'check_returncode': True,
+        }
+        lock = self.LOCK % node.hostname
+        cmd = '%s 2>&1' % ' '.join(update_cmd)
+
+        tid, err = execute(self.request, node.owner.id, cmd, meta=meta, lock=lock, queue=node.fast_queue,
+                           tg=TG_DC_UNBOUND)
+
+        if err:
+            return FailureTaskResponse(self.request, err, dc_bound=False)
+        else:
+            return TaskResponse(self.request, tid, msg=msg, obj=node, api_view=_apiview_, data=self.data,
+                                dc_bound=False, detail_dict=ser.detail_dict(force_full=True))
 
 
 class NodeLogsView(APIView):