+++ api/system/node/views.py
@@ -48,6 +48,8 @@ def system_node_version(request, hostname, data=None):
             * |async-no|
         :arg hostname: **required** - Node hostname
         :type hostname: string
+        :arg.data.fresh: Refresh cached node version information (default: false)
+        :type data.fresh: boolean
         :status 200: SUCCESS
         :status 403: Forbidden
         :status 404: Node not found
@@ -111,11 +113,11 @@ def system_node_service_status(request, hostname, name, data=None):
 @request_data_defaultdc(permissions=(IsSuperAdmin,))
 def system_node_update(request, hostname, data=None):
     """
-    Install (:http:put:`PUT </system/node/(hostname)/update>`) Danube Cloud update on a compute node.
+    Update (:http:put:`PUT </system/node/(hostname)/update>`) Danube Cloud on a compute node.
 
     .. http:put:: /system/node/(hostname)/update
 
-        .. note:: The compute node software will be updated to the \
+        .. note:: The compute node software should be updated to the \
 same :http:get:`version </system/version>` as installed on the main Danube Cloud management VM. \
 Use this API call after successful :http:put:`system update </system/update>`.
 
@@ -124,11 +126,13 @@ Use this API call after successful :http:put:`system update </system/update>`.
         :Permissions:
             * |SuperAdmin|
         :Asynchronous?:
-            * |async-no|
+            * |async-yes|
         :arg hostname: **required** - Node hostname
         :type hostname: string
         :arg data.version: **required** - git tag (e.g. ``v2.6.5``) or git commit to which the system should be updated
         :type data.version: string
+        :arg data.force: Whether to perform the update operation even though the software is already at selected version
+        :type data.force: boolean
         :arg data.key: X509 private key file used for authentication against EE git server. \
 Please note that file MUST contain standard x509 file BEGIN/END header/footer. \
 If not present, cached key file "update.key" will be used.
@@ -138,14 +142,12 @@ Please note that file MUST contain standard x509 file BEGIN/END headers/footer. 
 If not present, cached cert file "update.crt" will be used.
         :type data.cert: string
         :status 200: SUCCESS
+        :status 201: PENDING
         :status 400: FAILURE
         :status 403: Forbidden
         :status 404: Node not found
-        :status 417: Node update file is not available
-        :status 423: Node is not in maintenance state / Node version information could not be retrieved / \
-Task is already running
+        :status 423: Node is not in maintenance state / Node version information could not be retrieved
         :status 428: Node is already up-to-date
-        :status 504: Node worker is not responding
     """
     return NodeUpdateView(request, hostname, data).put()