+++ core/management/commands/gendoc.py
@@ -1,12 +1,13 @@
 import os
+import re
 
-from ._base import DanubeCloudCommand, CommandOption, lcd
+from ._base import DanubeCloudCommand, CommandOption, CommandError, lcd
 
 
 class Command(DanubeCloudCommand):
     help = 'Generate documentation files displayed in GUI.'
     DOC_REPO = 'https://github.com/erigones/esdc-docs.git'
-    DOC_TMP_DIR = '/tmp/esdc-docs'
+    DOC_TMP_DIR = '/var/tmp/esdc-docs'
     options = (
         CommandOption('--api', '--api-only', action='store_true', dest='api_only', default=False,
                       help='Generate only the API documentation.'),
@@ -48,28 +49,43 @@ class Command(DanubeCloudCommand):
 
         self.display('API documentation built successfully.', color='green')
 
-    def gendoc_user_guide(self):
+    def gendoc_user_guide(self, fallback_branch='master'):
         """Generate user guide"""
         doc_dst = self._path(self.PROJECT_DIR, 'gui', 'static', 'user-guide')
 
         with lcd(self.PROJECT_DIR):
-            branch = self.local('git rev-parse --abbrev-ref HEAD', capture=True).strip()
-            self.display('We are on branch "%s"' % branch)
+            try:
+                branch = self.get_git_version()[0]  # Git tag or branch name
+            except CommandError:
+                self.display('Could not determine our branch or tag', color='yellow')
+                branch = fallback_branch
+                self.display('Falling back to "%s" branch' % branch, color='yellow')
+            else:
+                self.display('We are on branch "%s"' % branch)
 
         if self._path_exists(self.DOC_TMP_DIR):
+            existing_repo = True
             self.display('%s already exists in %s' % (self.DOC_REPO, self.DOC_TMP_DIR), color='yellow')
             with lcd(self.DOC_TMP_DIR):
-                self.local('git pull')
+                self.local('git fetch')
             self.display('%s has been successfully updated.' % self.DOC_REPO, color='green')
         else:
+            existing_repo = False
             self.local('git clone %s %s' % (self.DOC_REPO, self.DOC_TMP_DIR))
             self.display('%s has been successfully cloned.' % self.DOC_TMP_DIR, color='green')
 
         with lcd(self.DOC_TMP_DIR):
-            if self.local('git checkout %s' % branch, raise_on_error=False) == 0:
-                self.display('Checked out esdc-docs branch "%s"' % branch, color='green')
-            else:
-                self.display('Could not checkout esdc-docs branch "%s"' % branch, color='red')
+            if self.local('git checkout %s' % branch, raise_on_error=False) != 0:
+                self.display('Could not checkout esdc-docs branch "%s"' % branch, color='yellow')
+                branch = fallback_branch
+                self.display('Falling back to "%s" branch' % branch, color='yellow')
+                self.local('git checkout %s' % branch)
+
+            self.display('Checked out esdc-docs branch "%s"' % branch, color='green')
+            # If the branch is no a tag name, then we need to merge/pull
+            if existing_repo and not re.search('^v[0-9]', branch):
+                self.local('git merge --ff-only')
+                self.display('Merged esdc-docs branch "%s"' % branch, color='green')
 
         # Build sphinx docs
         with lcd(self._path(self.DOC_TMP_DIR, 'user-guide')):